+++ docs/content/en/changelog.md
@@ -15,7 +15,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 **Added**
 
-- `toHaveJson()` expectation
+- `toBeRedirectToSignedRoute()` expectation
+- `toHaveExactJson()` expectation
 - `toHaveExactJson()` expectation
 
 ### [v1.3.0](https://github.com/pestphp/defstudio-plugin-laravel-expectations/compare/v1.2.0...v1.3.0) - 2021-15-10